requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==2.2.10
djangorestframework==3.13.1
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
